WIP: tou update

+++ source/Patches/MedicMod/Protect.cs
@@ -16,7 +16,7 @@ namespace TownOfUs.MedicMod
             if (PlayerControl.LocalPlayer.Data.IsDead) return false;
             if (role.UsedAbility) return false;
             var distBetweenPlayers = Utils.getDistBetweenPlayers(PlayerControl.LocalPlayer, role.ClosestPlayer);
-            var isCloseEnough = distBetweenPlayers < (double)GameOptionsData.KillDistances[PlayerControl.GameOptions.KillDistance];
+            var isCloseEnough = distBetweenPlayers < GameOptionsData.KillDistances[PlayerControl.GameOptions.KillDistance];
             if (!isCloseEnough) return false;
 
             var writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId,
